Session 62: Dtype compatibility fixes + baseline validation success
ACHIEVEMENT: First successful Q3-Omni 30B generation with SAGE! 🎉

Baseline Validation: ✅ COMPLETE (10/10 generations, perplexity 17.62)
Trust-Augmented: ⏸️  BLOCKED (dtype issue in attention layer)

Changes:
- Added NumPy import to selective_transformer_layer.py
- Fixed dtype handling in ContextClassifier (kept float64 for sklearn)
- Added float32 conversion in TrustBasedExpertSelector
- Added explicit dtype specs to tensor creation
- Added float32 conversion for expert weights

Remaining Issue:
- PyTorch C++ buffer dtype mismatch in attention computation
- Error occurs after projection, before MoE output
- Likely in attention mechanism internals
- Requires deeper investigation

Files Modified:
- sage/compression/selective_transformer_layer.py
- sage/core/context_classifier.py
- sage/core/trust_based_expert_selector.py

Status:
- Integration pathway (Phase 1-4): ✅ VALIDATED
- Production baseline: ✅ COMPLETE
- Trust-augmented production: ⏸️  INVESTIGATION NEEDED

+++ sage/core/trust_based_expert_selector.py
@@ -170,11 +170,11 @@ class TrustBasedExpertSelector:
                 # No context provided and no way to classify: use default
                 context = "general"
 
-        # Convert router logits to numpy for easier manipulation
+        # Convert router logits to numpy for easier manipulation (ensure float32)
         if HAS_TORCH and torch is not None and isinstance(router_logits, torch.Tensor):
-            router_scores = router_logits.detach().cpu().numpy()
+            router_scores = router_logits.detach().cpu().numpy().astype(np.float32)
         else:
-            router_scores = np.array(router_logits)
+            router_scores = np.array(router_logits, dtype=np.float32)
 
         # Get contextual trust for each expert
         trust_scores = self._get_contextual_trust_scores(context)
